+++ fujicv/models/__init__.py
@@ -4,6 +4,12 @@ from fujicv.models.backbone import build_backbone, list_available_backbones
 from fujicv.models.builder import ModelBuilder
 from fujicv.models.custom_layers import AttentionPool, GeM, LinearBNDropout, SqueezeExcite
 from fujicv.models.head import ClassificationHead, MultiLabelHead, RegressionHead
+from fujicv.models.metric_heads import (
+    AddMarginProduct,
+    ArcMarginProduct,
+    CosMarginProduct,
+    SubCenterArcMarginProduct,
+)
 
 __all__ = [
     "build_backbone",
@@ -16,4 +22,8 @@ __all__ = [
     "ClassificationHead",
     "RegressionHead",
     "MultiLabelHead",
+    "ArcMarginProduct",
+    "AddMarginProduct",
+    "CosMarginProduct",
+    "SubCenterArcMarginProduct",
 ]